requirements.txt
SQLAlchemy==1.4.32
psycopg2==2.9.3
google-api-python-client==2.41.0
gspread==5.2.0
oauthlib==3.2.0
python-telegram-bot==13.11
requests==2.27.1